Dei uma refinada na jogabilidade: - Score final do jogador - a correnteza vai aumentando durante o jogo - + efeitos e músicas

+++ graficos.h
@@ -34,7 +34,7 @@ int criaJanela(ALLEGRO_DISPLAY*, int, int);
 /*--6--*/
 void desenhaTeste(int);
 /*--7--*/
-int desenhaInfo(int, float, int);
+int desenhaInfo(int, int, int);
 
 
 /*-- 1 -- desenhaRio: Utiliza uma grade, ja com as velocidades calculadas
